+++ CHANGELOG.md
@@ -4,10 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+* Decrease horrendousness of performance for large projects containing large files
+
 ### API
 
 * IEnumerable<Task<ConversionResult>> becomes IAsyncEnumerable<ConversionResult>
 * Upgraded target framework from netstandard 1.3 to netstandard 2.0
+* Introduced cancellation token
 
 ### Vsix
 
@@ -25,7 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 * Exclude project file from conversion result if it hasn't changed
 * Further efforts to stop the roslyn library crashing Visual Studio
-* Improve support for large projects containing large files ()
+* Conversion tasks are now cancellable
 
 ### VB -> C#